pyproject.toml
[project]
name = "polyautomate"
version = "0.1.0"
description = "Utility clients for Polymarket data retrieval and trading automation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pynacl>=1.5.0",
]

[project.urls]
Homepage = "https://polymarket.com"

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
